Set up enzyme and jest and got one working test
